+++ nodes/knxUltimateHueButton.js
@@ -31,7 +31,12 @@ module.exports = function (RED) {
     node.setNodeStatus = ({ fill, shape, text, payload }) => {
 
     }
-
+    // Used to call the status update from the HUE config node.
+    node.setNodeStatusHue = ({ fill, shape, text }) => {
+      const dDate = new Date()
+      node.status({ fill: fill, shape: shape, text: text + ' (' + dDate.getDate() + ', ' + dDate.toLocaleTimeString() + ')' })
+    }
+    
     // This function is called by the knx-ultimate config node, to output a msg.payload.
     node.handleSend = msg => {
       const state = {}
@@ -158,6 +163,9 @@ module.exports = function (RED) {
     node.on('close', function (done) {
       if (node.server) {
         node.server.removeClient(node)
+      }
+      if (node.serverHue) {
+        node.serverHue.removeClient(node)
       }
       done()
     })