+++ nodes/knxUltimateHueTapDial.js
@@ -53,7 +53,12 @@ module.exports = function (RED) {
     node.setNodeStatus = ({ fill, shape, text, payload }) => {
 
     }
-
+    // Used to call the status update from the HUE config node.
+    node.setNodeStatusHue = ({ fill, shape, text }) => {
+      const dDate = new Date()
+      node.status({ fill: fill, shape: shape, text: text + ' (' + dDate.getDate() + ', ' + dDate.toLocaleTimeString() + ')' })
+    }
+    
     // This function is called by the knx-ultimate config node, to output a msg.payload.
     node.handleSend = msg => {
     }
@@ -88,7 +93,7 @@ module.exports = function (RED) {
               }
               // Send to KNX bus
               if (knxMsgPayload.topic !== '' && knxMsgPayload.topic !== undefined) node.server.writeQueueAdd({ grpaddr: knxMsgPayload.topic, payload: knxMsgPayload.payload, dpt: knxMsgPayload.dpt, outputtype: 'write', nodecallerid: node.id })
-              if (knxMsgPayload.topic !== '' && knxMsgPayload.topic !== undefined) node.status({ fill: 'green', shape: 'dot', text: 'HUE->KNX start Dim' + ' (' + new Date().getDate() + ', ' + new Date().toLocaleTimeString() + ')' })
+              if (knxMsgPayload.topic !== '' && knxMsgPayload.topic !== undefined) node.status({ fill: 'green', shape: 'dot', text: 'HUE->KNX Change color clockwise' + ' (' + new Date().getDate() + ', ' + new Date().toLocaleTimeString() + ')' })
             }
           } else if (_event.relative_rotary.last_event.rotation.direction === 'counter_clock_wise') {
             if (knxMsgPayload.dpt.startsWith('3.007')) {
@@ -112,7 +117,7 @@ module.exports = function (RED) {
               }
               // Send to KNX bus
               if (knxMsgPayload.topic !== '' && knxMsgPayload.topic !== undefined) node.server.writeQueueAdd({ grpaddr: knxMsgPayload.topic, payload: knxMsgPayload.payload, dpt: knxMsgPayload.dpt, outputtype: 'write', nodecallerid: node.id })
-              if (knxMsgPayload.topic !== '' && knxMsgPayload.topic !== undefined) node.status({ fill: 'green', shape: 'dot', text: 'HUE->KNX start Dim' + ' (' + new Date().getDate() + ', ' + new Date().toLocaleTimeString() + ')' })
+              if (knxMsgPayload.topic !== '' && knxMsgPayload.topic !== undefined) node.status({ fill: 'green', shape: 'dot', text: 'HUE->KNX Change color counterclockwise' + ' (' + new Date().getDate() + ', ' + new Date().toLocaleTimeString() + ')' })
             }
           }
 
@@ -158,6 +163,9 @@ module.exports = function (RED) {
     node.on('close', function (done) {
       if (node.server) {
         node.server.removeClient(node)
+      }
+      if (node.serverHue) {
+        node.serverHue.removeClient(node)
       }
       done()
     })